moved DummySpec, used to test sbt runner to examples project
